24/03: the three quality gates were built and never used anywhere
Nothing in the chapter ran check_coverage_gate, check_freshness_gate or
check_consistency_gate. They were defined inside 03_state_and_memory, exercised
on that notebook's own fixtures, and left there, while section 24.6 of the
chapter README describes abstention as part of the core project and the research
agent in 04 has no abstention path at all. Two claims 03 made about its
neighbours were false as a result.

The three functions and their date parser now live in agent_schemas.py beside
AgentState, which is what they operate on. 03 imports them, explains what each
one is arguing about, and keeps the four fixtures that show each condition
firing on its own. 04 builds an AgentState from the finished artifact's traces
and reports what the gates say about the pinned run.

Two of the three refuse it, and reporting that is the point:

- coverage fails because all four of the agent's searches asked what is
  happening now, so the forecast has no historical anchor. The system prompt
  never asks for one.
- consistency reports no cutoff to enforce. CHAPTER_CONTESTED_QUESTION carries
  cutoff_date="" because it was open when captured, so no evidence available
  that day could have held the answer. The gate having nothing to check is the
  correct outcome, and the notebook now says which questions it does bite on.

The second was previously asserted in prose only ("the trace cannot establish
point-in-time source availability"), which is the shape N3 calls the worst form:
a diagnostic the notebook promises and does not compute.

Two false statements 03 made, both found by roborev job 19279:

- It said 01_react_reasoning leaves no record when its conversation ends. 01
  writes a RunTrace holding every prompt, reply and search result, and replays
  from it. The section now distinguishes the trace, written from outside after
  the fact, from state the agent writes to during the run and a gate can read.
- It said 04 decides what to do on a gate failure. It did not.

Refs ml4t/agent-workspace#785
